requirements.txt
beautifulsoup4==4.11.1
selenium==4.6.1
webdriver-manager==3.8.5
python-dotenv==0.21.0
pandas==1.5.2
